+++ class/actes/ActesAnalyseFichierController.class.php
@@ -2,28 +2,28 @@
 
 class ActesAnalyseFichierController {
 
-    private $actes_files_upload_root;
     private $actes_appli_trigramme;
     private $actesTransactionsSQL;
     private $logger;
     private $actesEnvelopeSQL;
+    private $actesScriptHelper;
 
     public function __construct(
-        $actes_files_upload_root,
         Logger $logger,
         ActesTransactionsSQL $actesTransactionsSQL,
         ActesEnvelopeSQL $actesEnvelopeSQL,
-        $actes_appli_trigramme
+        $actes_appli_trigramme,
+        ActesScriptHelper $actesScriptHelper
     ) {
-        $this->actes_files_upload_root = $actes_files_upload_root;
         $this->actes_appli_trigramme = $actes_appli_trigramme;
         $this->logger = $logger;
         $this->actesTransactionsSQL = $actesTransactionsSQL;
         $this->actesEnvelopeSQL = $actesEnvelopeSQL;
+        $this->actesScriptHelper = $actesScriptHelper;
     }
 
     private function log($message){
-        $this->logger->log("Actes analyse fichier à envoyer",$message);
+        $this->logger->log("actes-analyse-fichier-a-envoyer",$message);
     }
 
     public function validateAllEnveloppe(){
@@ -33,7 +33,6 @@ class ActesAnalyseFichierController {
         foreach($enveloppe_ids as $enveloppe_id){
             $this->validateOneEnveloppe($enveloppe_id);
         }
-
         $this->log("Fin du script");
         return true;
     }
@@ -45,9 +44,7 @@ class ActesAnalyseFichierController {
 
         $this->log("[$envelope_libelle] Analyse");
 
-        $envelope_info = $this->actesEnvelopeSQL->getInfo($enveloppe_id);
-
-        $archive_path =  $this->actes_files_upload_root . "/" . $envelope_info['file_path'];
+        $archive_path =  $this->actesScriptHelper->getArchivePath($enveloppe_id);
 
         $this->log("[$envelope_libelle] Emplacement de l'archive :  $archive_path");
 
@@ -57,40 +54,17 @@ class ActesAnalyseFichierController {
         } catch (Exception $e){
             $message = utf8_decode( $e->getMessage());
             $this->log("[$envelope_libelle] L'archive n'est valide : $message");
-            $this->updateStatus($transaction_ids,ActesStatusSQL::STATUS_EN_ERREUR,"Enveloppe invalide : $message");
+            $this->actesScriptHelper->updateStatus($transaction_ids,ActesStatusSQL::STATUS_EN_ERREUR,"Enveloppe invalide : $message");
             return false;
         }
         $this->log("[$envelope_libelle] L'archive est valide !");
-        $this->updateStatus(
+        $this->actesScriptHelper->updateStatus(
             $transaction_ids,
             ActesStatusSQL::STATUS_EN_ATTENTE_DE_TRANSMISSION,
             "Accepte par le TdT : validation OK"
         );
 
         return true;
-    }
-
-    private function updateStatus($transactions_ids,$status_id,$message,$flux_retour = ""){
-        foreach($transactions_ids as $transactions_id) {
-            $this->actesTransactionsSQL->updateStatus($transactions_id,$status_id,$message,$flux_retour);
-            $info = $this->actesTransactionsSQL->getInfo($transactions_id);
-            //WTF : trouvé dans le code java... j'imagine que ca sert pour l'API ?
-            $message_log = "L'archive $transactions_id passe à l'état " . $this->getStatusName($status_id);
-
-            if ($status_id == ActesStatusSQL::STATUS_ACQUITTEMENT_RECU) {
-                $message_log = "L'acte : $transactions_id passe en recu.";
-            }
-			if ($status_id == ActesStatusSQL::STATUS_ANNULER) {
-                $message_log = "L'acte : $transactions_id passe en annulee.";
-            }
-            Log::newEntry(LOG_ISSUER_NAME,$message_log,1,false,"USER","actes",false,$info['user_id']);
-        }
-    }
-    public function getStatusName($status_id){
-        $status_list = array(
-            -1 => 'erreur', 'annule', 'poste', 'en attente','transmis','recu'
-        );
-        return $status_list[$status_id];
     }
 
 }