+++ class/actes/ActesFileSender.class.php
@@ -1,0 +1,42 @@
+<?php
+
+class ActesFileSender {
+
+    private $actesMinistereProperties;
+
+    public function __construct(
+        ActesMinistereProperties $actesMinistereProperties
+    ) {
+        $this->actesMinistereProperties = $actesMinistereProperties;
+    }
+
+    public function send($filepath){
+       $curlWrapper = new CurlWrapper();
+
+        $url = $this->actesMinistereProperties->url;
+
+        if(  $this->actesMinistereProperties->authentification_type == ActesMinistereProperties::AUTHENTICATION_POST){
+            $url .= "?user={$this->actesMinistereProperties->login}&password={$this->actesMinistereProperties->password}";
+        }
+        if ($this->actesMinistereProperties->authentification_type == ActesMinistereProperties::AUTHENTICATION_BASIC){
+            $curlWrapper->httpAuthentication($this->actesMinistereProperties->login,$this->actesMinistereProperties->password);
+        }
+
+        $curlWrapper->setClientCertificate(
+            $this->actesMinistereProperties->client_certificate,
+            $this->actesMinistereProperties->client_certificate_key,
+            $this->actesMinistereProperties->client_certificate_key_password
+        );
+
+        $curlWrapper->addPostFile(basename($filepath),$filepath);
+
+        $curlWrapper->get($url);
+
+        if ($curlWrapper->getHTTPCode() != 200){
+            throw new Exception($curlWrapper->getLastError());
+        }
+
+        return true;
+    }
+
+}