+++ CHANGELOG.md
@@ -9,6 +9,8 @@ et adhère aux principes du [Semantic Versioning](http://semver.org/).
 
 ### Ajouts
 - Création du script cron/actes-analyse-fichier-a-envoyer.php qui vise à remplacer la partie Tomcat #190 ~Actes
+- Création du script cron/acte-envoi-fichier.php pour remplacer la partie Tomcat #192 ~Actes
+  
 
 ### Corrections de sécurités